+++ docs/source/changelog.rst
@@ -13,6 +13,7 @@ Dev branch
   definitions to the Hammer Addons repository.
 * Added new :py:meth:`srctools.fgd.EntityDef.get_resources()` method, replacing ``fgd.entclass_*()``
   methods.
+* When parsing VMF outputs, assume extraneous commas are part of the parameter.
 
 -------------
 Version 2.3.5